backend/requirements.txt
Django>=4.0,<4.1
djangorestframework>=3.13,<3.14
django-cors-headers>=3.13,<3.14
djangorestframework_simplejwt>=5.2,<5.3
social-auth-app-django>=4.0,<4.1
drf-spectacular>=0.24.2,<0.25
drf-yasg>=1.21.4,<1.22
Pillow>=9.2.0,<9.3
psycopg2>=2.9.3,<2.10
uwsgi>=2.0.20<2.1